Relocate the pipeline examples
