+++ pipeline/example/custom/indexing/fetcher_tasks.go
