+++ pipeline/example/custom/indexing/payload.go
