+++ pipeline/example/custom/indexing/pipeline.go
