+++ pipeline/example/custom/indexing/sink.go
