+++ pipeline/example/custom/indexing/source.go
