+++ pipeline/example/custom/indexing/syncer_tasks.go
