+++ pipeline/example/custom/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/figment-networks/indexing-engine/example/default/indexing"
+	"github.com/figment-networks/indexing-engine/pipeline/example/custom/indexing"
 )
 
 func main() {
