+++ pipeline/example/default/indexing/aggregator_tasks.go
