+++ pipeline/example/default/indexing/fetcher_tasks.go
