+++ pipeline/example/default/indexing/payload.go
