+++ pipeline/example/default/indexing/pipeline.go
