+++ pipeline/example/default/indexing/sink.go
