+++ pipeline/example/default/indexing/source.go
