+++ pipeline/example/default/indexing/syncer_tasks.go
