+++ pipeline/example/default/indexing/validator_tasks.go
